feat: Assistant 및 TextToSpeech 서비스 인터페이스 추가 및 기존 서비스 리팩토링
- AssistantService를 Assistant 인터페이스로 변경하고, 관련 메서드 구현
- TextToSpeechService를 TextToSpeech 인터페이스로 변경하고, 관련 메서드 구현
- AssistantController 및 TextToSpeechController에서 서비스 의존성 변경
- LlmProxyService 삭제 및 관련 코드 정리
