+++ src/main/java/com/newworld/saegil/llm/service/Assistant.java
@@ -1,0 +1,19 @@
+package com.newworld.saegil.llm.service;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface Assistant {
+
+    AssistantResponse getAssistantTextResponseFromAudioFile(
+            MultipartFile multipartFile,
+            String threadId
+    );
+
+    AssistantResponse getAssistantTextResponseFromAudioFile(
+            MultipartFile multipartFile,
+            String threadId,
+            Long scenarioId,
+            Long userId
+    );
+
+}
