+++ src/main/java/com/newworld/saegil/llm/service/TextToSpeech.java
@@ -1,0 +1,10 @@
+package com.newworld.saegil.llm.service;
+
+import com.newworld.saegil.llm.config.TtsProvider;
+import org.springframework.core.io.Resource;
+
+public interface TextToSpeech {
+
+    Resource convertTextToSpeech(final String text, final TtsProvider provider);
+
+}
